Add encryption layer via TLS
With plain TCP, anyone MITM is able to read the data exchanged between
the two parts. Adding TLS prevents this without preventing MITM attacks
entirely.
The next step is to add some kind of authentication mechanism so the
user can make sure it is sending/receiving the data from the right host.